requirements.txt
sentence-transformers
sickit-learn
umap-learn
matplotlib